fix: FR-265 Remove additional parameters in index controller

See merge request magesuite/brand-management!72

+++ Test/Integration/_files/brand_with_products_rollback.php
@@ -3,19 +3,32 @@ $objectManager = \Magento\TestFramework\Helper\Bootstrap::getObjectManager();
 $registry = $objectManager->get(\Magento\Framework\Registry::class);
 $productRepository = $objectManager->create(\Magento\Catalog\Api\ProductRepositoryInterface::class);
 $brandRepository = $objectManager->create(\MageSuite\BrandManagement\Api\BrandsRepositoryInterface::class);
+$installer = $objectManager->create(\Magento\Catalog\Setup\CategorySetup::class);
 
-$productSkus = ['sku111', 'sku112', 'sku113', 'sku114', 'sku115'];
+$productSkus = array_map(
+    function ($number) {
+        return sprintf('sku%s', $number);
+    },
+    range(111, 120)
+);
 
 $registry->unregister('isSecureArea');
 $registry->register('isSecureArea', true);
 
 foreach ($productSkus as $sku) {
-    $product = $productRepository->get($sku, false, null, true);
-    $productRepository->delete($product);
+    try {
+        $product = $productRepository->get($sku);
+        $productRepository->delete($product);
+    } catch (\Exception $e) {
+        continue;
+    }
 }
 
 $brand = $brandRepository->getById(111);
-$brand->delete();
+
+if ($brand) {
+    $brand->delete();
+}
 
 $registry->unregister('isSecureArea');
 $registry->register('isSecureArea', false);
